refactor: split executor_files.go by concern
